AudioManager, volume slider, pause menu
Added an audio manager, which controls all the volume and allows setting a master volume for the whole game without using an AudioMix, meaning it should work in the WebGL build. The game can now be paused. Volume can be controlled from the pause menu. The player can also return to the main menu from the pause menu, and changes to the volume carry over between scenes and between sessions.